requirements.txt
pynput=1.7.6
pyperclip=1.8.2
request=2.31.0